Jd global changes

+++ src/app/api/admin/deleteproduct/route.ts
@@ -1,10 +1,22 @@
 import prisma from "@/services/db";
+import admin from "@/services/firebase-admin-config";
 import { db } from "@/services/firebase-config";
 import { addDoc, collection } from "firebase/firestore";
 import { NextRequest } from "next/server";
 
 export async function PUT(req: NextRequest) {
     const { data } = await req.json();
+
+    const decodedToken = await admin.auth().verifyIdToken(data.token);
+
+    if (!decodedToken) {
+        return new Response(JSON.stringify("Unauthorized"), {
+            status: 403,
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        });
+    };
 
         // Check if the item already has an active request with itemStatusId: 3
         const result = await prisma.$transaction(async (prisma) => {
@@ -25,6 +37,24 @@ export async function PUT(req: NextRequest) {
                     role: true,
                 }
             });
+
+            if (!user) {
+                return new Response(JSON.stringify("User not found"), {
+                    status: 404,
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+            };
+        
+            if (!["Admin"].includes(user.role.name)) {
+                return new Response(JSON.stringify("Forbidden, you don't have the rights to make this call"), {
+                    status: 403, // Use 403 for Forbidden instead of 404
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+            };
     
             const Admin = await prisma.user.findMany({
                 where: {
@@ -68,6 +98,17 @@ export async function PUT(req: NextRequest) {
 export async function DELETE(req: NextRequest) {
     const { data } = await req.json();
 
+    const decodedToken = await admin.auth().verifyIdToken(data.token);
+
+    if (!decodedToken) {
+        return new Response(JSON.stringify("Unauthorized"), {
+            status: 403,
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        });
+    };
+
         // Check if the item already has an active request with itemStatusId: 3
         const result = await prisma.$transaction(async (prisma) => {
             await prisma.roleItem.deleteMany({
@@ -90,6 +131,24 @@ export async function DELETE(req: NextRequest) {
                     role: true,
                 }
             });
+
+            if (!user) {
+                return new Response(JSON.stringify("User not found"), {
+                    status: 404,
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+            };
+        
+            if (!["Admin"].includes(user.role.name)) {
+                return new Response(JSON.stringify("Forbidden, you don't have the rights to make this call"), {
+                    status: 403, // Use 403 for Forbidden instead of 404
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+            };
     
             const Admin = await prisma.user.findMany({
                 where: {
